+++ client/src/components/CreateItem.js
@@ -1,31 +1,43 @@
 import React, { useState, useEffect } from 'react';
-import Axios from 'axios';
 import { Form, Button } from 'react-bootstrap';
 import { Link } from 'react-router-dom';
 import { useAuth0 } from '@auth0/auth0-react';
 import TabelItems from './TableItems';
+import io from 'socket.io-client';
 
 export default function CreateItem(props) {
+    const socket = io();
     const [collection, setCollection] = useState([]);
     const [itemData, setItemData] = useState({});
-    const { user } = useAuth0();
+    const [dataItems, setDataItems] = useState([]);
+    const { user, isAuthenticated } = useAuth0();
 
     useEffect(()=>{
-        Axios.post('/getCollectionInfo',{
+        socket.emit('getCollectionInfo',{
             _id: props.location.pathname.slice(12)
         })
-        .then((res) => setCollection(res.data))
-        .catch((e) => console.log(e))
+        socket.on('getCollectionDataInfo', (data) => {
+            setCollection(data)
+        })
+        socket.emit('getItems', {
+            idCollect: (isAuthenticated) ? props.location.pathname.slice(12) : 'all'
+        })
+        socket.on('getDataItems',(data) => {
+            setDataItems(data);
+        })
     },[])
 
     const addItem = (e) => {
         e.preventDefault();
-        Axios.post('/addItem', {
+        socket.emit('addItem', {
             idUser: user.sub,
             idCollect: props.location.pathname.slice(12),
             dataItem: JSON.stringify(itemData),
             poleItem: collection[0].poleItem
         })
+        socket.on('getDataItems',(data) => {
+            setDataItems(data);
+        },[])
     }
 
     return (
@@ -37,16 +49,30 @@ export default function CreateItem(props) {
                 <Form>
                     {(collection.length != 0)&&(
                         Object.keys(JSON.parse(collection[0].poleItem)).map((keyName, idx) => {
-                            if (JSON.parse(collection[0].poleItem)[keyName] != 'textarea' ) {
+                            if (JSON.parse(collection[0].poleItem)[keyName] == 'checkbox' ) {
+                                return(
+                                    <Form.Check type="checkbox" id="autoSizingCheck2" label={keyName} />
+                                )
+                            } else if(JSON.parse(collection[0].poleItem)[keyName] == 'textarea') {
                                 return (
-                                   <Form.Group key={idx} controlId={`ControlInput${idx}`}>
+                                <Form.Group key={idx} controlId={`ControlInput${idx}`}>
                                        <Form.Label>Введите {`${keyName}`}: </Form.Label>
-                                       <Form.Control type={`${JSON.parse(collection[0].poleItem)[keyName]}`} title={keyName} onChange={(e)=>{
+                                       <Form.Control as={`${JSON.parse(collection[0].poleItem)[keyName]}`} rows={3} title={keyName} onChange={(e)=>{
                                            itemData[e.target.title] = e.target.value;
                                            setItemData(itemData);
                                        }} />
-                                   </Form.Group>
+                                </Form.Group>
                                 )
+                            } else {
+                                return (
+                                    <Form.Group key={idx} controlId={`ControlInput${idx}`}>
+                                        <Form.Label>Введите {`${keyName}`}: </Form.Label>
+                                        <Form.Control type={`${JSON.parse(collection[0].poleItem)[keyName]}`} title={keyName} onChange={(e)=>{
+                                            itemData[e.target.title] = e.target.value;
+                                            setItemData(itemData);
+                                        }} />
+                                    </Form.Group>
+                                 )
                             }
                         })
                     )}
@@ -56,7 +82,7 @@ export default function CreateItem(props) {
                 </Form>
             </div>
         </div>
-        <TabelItems idCollect={props.location.pathname.slice(12)} />  
+        <TabelItems dataItems={dataItems} idCollect={props.location.pathname.slice(12)} />  
     </div>
     )
 }