+++ server.js
@@ -5,6 +5,11 @@ const bodyParser = require('body-parser');
 const http = require('http').createServer(app);
 const path = require('path');
 const cors = require('cors');
+const io = require('socket.io')(http, {
+    cors: {
+      origin: '*',
+    }
+  });
 const PORT = process.env.PORT || 3001;
 
 const collections = require('./config/collectionSchema');
@@ -37,71 +42,110 @@ mongoose
     .then(() => console.log('MongoDb connected'))
     .catch((e) => console.log(e))
 
-app.post('/addCollection', (req,res) =>{
-    const { id, name, description, teg, url, poleItem } = req.body;
-    collections.create({
-        id: id,
-        name: name,
-        description: description,
-        teg: teg,
-        url: url,
-        poleItem: poleItem
-    })
-    .catch((err)=> console.log(err))
-})
 
-app.post('/addItem', (req,res) =>{
-    const { idUser, idCollect, dataItem, poleItem } = req.body;
-    items.create({
-        idUser: idUser,
-        idCollect: idCollect,
-        dataItem: dataItem,
-        poleItem: poleItem
-    })
-    .catch((err)=> console.log(err))
-})
 
-app.post('/getItems', (req,res) =>{
-    const { idCollect } = req.body;
-    if (idCollect == 'all') {
-        items.find()
-        .then((data) => res.send(data))
+io.on("connection", function(socket) {
+    socket.on('addItem', async (data)=>{
+        const { idUser, idCollect, dataItem, poleItem } = data;
+        await items.create({
+            idUser: idUser,
+            idCollect: idCollect,
+            dataItem: dataItem,
+            poleItem: poleItem
+        })
         .catch((err)=> console.log(err))
-    } else {
-        items.find({ idCollect: idCollect })
-        .then((data) => res.send(data))
+
+        await items.find({ idCollect: idCollect })
+            .then((data) => socket.emit('getDataItems', data))
+            .catch((err)=> console.log(err))
+    })
+
+    socket.on('getItems', (data) => {
+        const { idCollect } = data;
+        if (idCollect == 'all') {
+            items.find()
+            .then((data) => socket.emit('getDataItems', data))
+            .catch((err)=> console.log(err))
+        } else {
+            items.find({ idCollect: idCollect })
+            .then((data) => socket.emit('getDataItems', data))
+            .catch((err)=> console.log(err))
+        }
+    })
+
+    socket.on('getCollectionInfo', (data) => {
+        const { _id } = data;
+        collections.find({
+            _id: _id
+        })
+        .then((data) =>{
+            return socket.emit('getCollectionDataInfo', data);
+        })
+        .catch((err) => console.log(err))
+    })
+
+    socket.on('getCollection', (data) => {
+        const { id } = data;
+        collections.find({
+            id: id
+        })
+        .then((data) =>{
+            return socket.emit('getDataCollect', data);
+        })
+        .catch((err) => console.log(err))
+    })
+
+    socket.on('addCollection', async (data) =>{
+        const { id, name, description, teg, url, poleItem } = data;
+        await collections.create({
+            id: id,
+            name: name,
+            description: description,
+            teg: teg,
+            url: url,
+            poleItem: poleItem
+        })
         .catch((err)=> console.log(err))
-    }
-})
-app.post('/getItemInfo', (req,res) => {
-    const { _id } = req.body;
-    items.find({
-        _id: _id
+        await collections.find({
+            id: id
+        })
+        .then((data) =>{
+            return socket.emit('getDataCollect', data);
+        })
+        .catch((err) => console.log(err))
     })
-    .then((data) =>{
-        return res.send(data);
-    })
-    .catch((err) => console.log(err))
-})
 
-app.post('/getCollectionInfo', (req,res) => {
-    const { _id } = req.body;
-    collections.find({
-        _id: _id
+    socket.on('getItemInfo', (data) => {
+        const { _id } = data;
+        items.find({
+            _id: _id
+        })
+        .then((data) =>{
+            return socket.emit('getItemDataInfo', data);
+        })
+        .catch((err) => console.log(err))
     })
-    .then((data) =>{
-        return res.send(data);
-    })
-    .catch((err) => console.log(err))
-})
 
-app.post('/getCollection', (req,res) => {
-    const { id } = req.body;
-    collections.find({
-        id: id
+    socket.on('deleteItem', (data) => {
+        const { _id } = data;
+        items.deleteOne({
+            _id: _id
+        })
+        .catch((err) => console.log(err))
     })
-    .then((data) =>{
-        return res.send(data);
+
+    socket.on('editItem', (data) =>{
+        const { _id, dataItem } = data;
+        items.updateOne(
+        {
+            _id: _id
+        },
+        {
+            $set: {dataItem: dataItem}
+        },
+        (err, result) => {
+            if(err) console.log(err);
+        }
+        )
     })
-    .catch((err) => console.log(err))
 })