+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on forthetraveler.netlify.app(https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2021-01-05
+### Edit
+
+- Communication with the server organizing through socket.io.
+
+### Added
+
+- Added the ability to update and delete item.
+
 ## [1.4.0] - 2021-01-04
 ### Edit